tests/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
docker>=6.1.0
pyyaml>=6.0
